Move ExpansionOptions enum on Enums folder
